use short instead of int in for loop
as the value is used to lookup in an array with the dimension of a short
a value greater than a short does not make sense.

+++ jsettlers.logic/src/main/java/jsettlers/algorithms/fogofwar/FogOfWar.java
@@ -126,8 +126,8 @@ public final class FogOfWar implements Serializable {
 		private void rebuildSight() {
 			drawer.setBuffer(buffer);
 
-			for (int x = 0; x < width; x++) {
-				for (int y = 0; y < height; y++) {
+			for (short x = 0; x < width; x++) {
+				for (short y = 0; y < height; y++) {
 					byte currSight = sight[x][y];
 
 					if (currSight >= CommonConstants.FOG_OF_WAR_EXPLORED) {
